requirements.txt
# PyTorch - GPU 서버용 (CUDA 12.1 기준)
--extra-index-url https://download.pytorch.org/whl/cu121

# ==========================================
# Web Server & Data Utils
# ==========================================
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
websockets>=12.0
requests>=2.32.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
numpy>=1.26.0


# ==========================================
# OpenAI API (GPT-4o-mini 플랜 생성용)
# ==========================================
openai>=1.0.0

# ==========================================
# [1] LLM (대형 언어 모델 - Qwen 등)
# ==========================================
transformers>=4.45.0
accelerate>=1.0.0
peft>=0.13.0
bitsandbytes>=0.43.0
protobuf>=5.0.0
sentencepiece>=0.2.0

# ==========================================
# [2] 얼굴 감정 인식 (Facial Emotion - DeepFace 등)
# ==========================================
deepface>=0.0.93
opencv-python-headless>=4.10.0
tf-keras>=2.15.0

# ==========================================
# [3] 오디오 인식 및 처리 (Audio/Speech - Whisper 등)
# ==========================================
faster-whisper>=1.0.0
soundfile>=0.12.0
librosa>=0.10.0
torchaudio>=2.3.0